Better support of HTTP nature #7
ackInput step is thought to read headers/contents of the request and set the response headers accordingly giving an acknowledgment of the received request (ie request has been accepted and consequences of such a request are guaranteed.

processOutput just build the response output according to the request. This may fail but client will not now it until it check the contentLength or TCP  connection errors. This is mainly not a problem for server integrity.