refactor(@whook/aws-lambda): refactor lambda triggers
Refactoring the wrappers to simplify them except for tricky parts like encoding/decoding of events.

BREAKING CHANGE: The consumers do not manage batchs anymore and simply pass the records to the
lambda implementation in order to avoid having to write custom wrappers for each event type. The
user are now free to handle the batchs the way they want.

fix #95 concerns #96