feat(@whook/cors): add an error wrapper for CORS
When an error was thrown at the router level, it had no CORS due to the fact that it did not hit any
handler and subsequently nor the handler CORS wrapper. This commit add a way to override the
errorHandler of a Whook server for both defaut and custom handlers.